% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grams.R
\name{alcohol_g_ethanol}
\alias{alcohol_g_ethanol}
\title{alcohol_g_ethanol
Function to compute g of alcohol per day from ml}
\usage{
alcohol_g_ethanol(
  grams_df,
  alcohol_items = c("beer", "red_wine", "white_wine", "liqueur_schnapps"),
  abv = c(beer = 0.05, red_wine = 0.12, white_wine = 0.12, liqueur_schnapps = 0.35),
  rho = 0.789,
  suffix = "__ethanol_g"
)
}
\arguments{
\item{grams_df}{grams dataframe}

\item{alcohol_items}{which alcohol items exist}

\item{abv}{approximate percentage}

\item{rho}{ethanol density}

\item{suffix}{which suffix should be given to the new column?}
}
\value{
dataframe with appended ethanol g per item and overall
}
\description{
alcohol_g_ethanol
Function to compute g of alcohol per day from ml
}
