% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grams.R
\name{grams}
\alias{grams}
\title{Grams}
\usage{
grams(
  df,
  label_col = "varlabel",
  name_col = "foodname",
  serving_col = "serving_size",
  unit_col = "unit"
)
}
\arguments{
\item{df}{A dataframe representing different frequency factor
responses from the FFQ (columns = items, rows = individuals)}

\item{label_col}{Column name of column containing food label variable}

\item{name_col}{Column name of column containing food name}

\item{serving_col}{Column name of column containing serving size}

\item{unit_col}{Column name of column containing serving size unit}
}
\value{
A dataframe representing the estimated total
daily grams (or ml) of each food consumed for the participant.
}
\description{
The grams function takes responses on the eutops FFQ to generate an
estimate of the participant's total daily units/grams consumed
for each food.
}
