% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{hPDI}
\alias{hPDI}
\title{healthy Plant Diet Index}
\usage{
hPDI(components)
}
\arguments{
\item{components}{output of make_components()}
}
\value{
healthy plant diet index, ranging from ~ -9 to +3
}
\description{
healthy Plant Diet Index
}
