% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{LCD}
\alias{LCD}
\title{Low carb diet score}
\usage{
LCD(summary)
}
\arguments{
\item{summary}{nutrients() summary of carb/protein/fat per day}
}
\value{
Low-carb diet score in percentiles, whigh higher scores being lower carb
}
\description{
Low carb diet score
}
