% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro_distribution.R
\name{macro_distribution}
\alias{macro_distribution}
\title{Title}
\usage{
macro_distribution(summary_df)
}
\arguments{
\item{summary_df}{Outputs of nutrients}
}
\value{
macro distribution dataframe
}
\description{
Title
}
