% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{make_components}
\alias{make_components}
\title{make components for dietary indices}
\usage{
make_components(intake)
}
\arguments{
\item{intake}{output of grams function}
}
\value{
dataframe with components grouped
}
\description{
make components for dietary indices
}
