% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{mds_trichopoulou}
\alias{mds_trichopoulou}
\title{Mediterranean Diet score (MDS) by Trichopoulou}
\usage{
mds_trichopoulou(
  components,
  alcohol_col = "alcohol",
  alcohol_range_g = c(5, 25),
  include_alcohol = TRUE
)
}
\arguments{
\item{components}{output of make_components from grams function}

\item{alcohol_col}{label of alcohol column}

\item{alcohol_range_g}{range - 5 to 25}

\item{include_alcohol}{T or F}
}
\value{
MDS score tibble
}
\description{
Mediterranean Diet score (MDS) by Trichopoulou
}
