% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{modifiedDASH}
\alias{modifiedDASH}
\title{modifiedDASH (7) for DASH diet index}
\usage{
modifiedDASH(components, ssb_col = "ssb")
}
\arguments{
\item{components}{output of make_components() from grams()}

\item{ssb_col}{soda column label}
}
\value{
dash score DF
}
\description{
modifiedDASH (7) for DASH diet index
}
