% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrients.R
\name{nutrients}
\alias{nutrients}
\title{Nutrients}
\usage{
nutrients(
  df,
  label_col = "varlabel",
  name_col = "foodname",
  nutrient_cols = c("serving_size", "calories_kcal", "protein_g", "fat_g", "carb_g",
    "calcium_mg", "iron_mg", "zinc_mg", "vitamin_c_mg", "vitamin_b6_mg", "vitamin_a_mg",
    "folate_mg")
)
}
\arguments{
\item{df}{data frame with rows (individual observations) and
columns, representing 90 different frequency factor responses from the
eutops FFQ}

\item{label_col}{Column name of column containing food label variable}

\item{name_col}{Column name of column containing food name}

\item{nutrient_cols}{Column names of nutrient outputs to be returned}
}
\value{
a list with two items: summary and per food group nutrients and
macros and per_food group
}
\description{
The Nutrients function takes the age of a participant and their responses on
the hsffq to generate an estimate of the participant's total daily
micronutrients, macronutrients, and calories consumed for each food
}
