% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{pick_items}
\alias{pick_items}
\title{pick_items is a helper function for make_components to ensure we can grab
the relevant items (within make_components)}
\usage{
pick_items(df, items, unit = c("g", "ml", "ethanol_g"))
}
\arguments{
\item{df}{output of the grams() function element}

\item{items}{this is defined in the make_components function}

\item{unit}{unit element}
}
\value{
rowSums
}
\description{
pick_items is a helper function for make_components to ensure we can grab
the relevant items (within make_components)
}
