% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotient.R
\name{quotient}
\alias{quotient}
\title{Food Quotient Based on Macronutrients computed from nutrients()}
\usage{
quotient(df, protein_col = "protein_g", fat_col = "fat_g", carb_col = "carb_g")
}
\arguments{
\item{df}{data frame (returned from nutrients() function)}

\item{protein_col}{Column name of column containing protein_g information}

\item{fat_col}{Column name of column containing fat_g information}

\item{carb_col}{Column name of column containing carb_g information}
}
\value{
one value per participant will be returned, representing the food
  quotient for the individual
}
\description{
The macquotient function calculates a food quotient for a participant based
on average daily protein, carbs, and fat consumed for an individual or a
group. In contrast to the quotient function, macquotient is able to generate
reliable average food quotients for a group of people rather than only
individual level. Group level estimates are recomended in some studies to
control for response bias.
}
