% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{RefCarb}
\alias{RefCarb}
\title{Refined carb score}
\usage{
RefCarb(components)
}
\arguments{
\item{components}{output of make_components()}
}
\value{
refined carb score value; quintiles from 1-5 (lower is better)
}
\description{
Refined carb score
}
